requirements.txt
# Core ML and Data Science packages
torch>=2.0.0
torchvision>=0.15.0
xgboost>=1.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=9.0.0

# Image processing
opencv-python>=4.8.0

# Web framework (if you're building a web API)
fastapi>=0.100.0
uvicorn>=0.23.0

# Utility packages
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Medical imaging (if needed)
pydicom>=2.4.0
nibabel>=5.1.0
